pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "fin-cli"
version = "0.1.0-dev1"
description = "Composable local-first financial CLI toolkit."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "GiantThings" },
]
license = { file = "LICENSE" }
keywords = ["finance", "cli", "personal-finance", "analysis"]
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Topic :: Office/Business :: Financial :: Accounting",
]
dependencies = [
  "click>=8.1",
  "rich>=13.7",
  "pyyaml>=6.0",
  "python-dateutil>=2.9",
  "pandas>=2.2",
  "jinja2>=3.1",
]

[project.optional-dependencies]
pdf = [
  "pdfplumber>=0.11",
  "camelot-py[base]>=0.11",
]
llm = [
  "openai>=1.40",
]
analysis = [
  "numpy>=1.26",
]
pii = [
  "scrubadub>=2.0",
]
dev = [
  "pytest>=8.2",
  "pytest-mock>=3.12",
  "ruff>=0.6",
  "black>=24.4",
  "mypy>=1.11",
]
all = [
  "pdfplumber>=0.11",
  "camelot-py[base]>=0.11",
  "openai>=1.40",
  "numpy>=1.26",
  "scrubadub>=2.0",
]

[project.scripts]
fin-query = "fin_cli.fin_query.main:main"
fin-analyze = "fin_cli.fin_analyze.main:main"
fin-edit = "fin_cli.fin_edit.main:main"
fin-scrub = "fin_cli.fin_scrub.main:main"

[tool.hatch.build.targets.sdist]
include = ["fin_cli", "README.md", "specs", "plans", "AGENTS.md"]

[tool.hatch.build.targets.wheel]
packages = ["fin_cli"]
include = [
  "fin_cli/fin_extract/bundled_specs/*.yaml",
  "fin_cli/fin_scrub/default_config.yaml",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "NPY", "G"]
ignore = [
    "E501",  # Line too long (handled by black)
    "G004",  # Logging f-string (common and acceptable in modern Python)
    "G003",  # Logging with + (acceptable alternative to %)
    "B904",  # raise-without-from-inside-except (too strict for many cases)
    "B905",  # zip-without-explicit-strict (Python 3.10+ feature, may not always be needed)
    "E402",  # Module level import not at top (intentional in some CLI entry points)
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_return_any = true
ignore_missing_imports = true
packages = ["fin_cli"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
filterwarnings = [
  "error::DeprecationWarning",
]
testpaths = ["tests"]
